refactor: extract board notifications module
