+++ packages/board/server.mjs
@@ -54,6 +54,7 @@ import {
   resolveProjectInfo,
 } from './lib/projects.mjs';
 import { broadcast } from './lib/sse.mjs';
+import { loadNotifHistory, saveNotifHistory, addNotification } from './lib/notifications.mjs';
 
 // ── Helpers ────────────────────────────────────────────────────────────────────
 
@@ -65,48 +66,6 @@ function broadcastTasks(cwd) {
     }
   }
 }
-
-function loadNotifHistory() {
-  try {
-    if (fs.existsSync(NOTIF_HISTORY_FILE)) {
-      const parsed = JSON.parse(fs.readFileSync(NOTIF_HISTORY_FILE, 'utf8'));
-      if (Array.isArray(parsed)) { notifHistory.length = 0; notifHistory.push(...parsed); }
-    }
-  } catch { /* start fresh on corrupt file */ }
-}
-
-function saveNotifHistory() {
-  try {
-    fs.mkdirSync(GREAT_CTO_DIR, { recursive: true });
-    fs.writeFileSync(NOTIF_HISTORY_FILE, JSON.stringify(notifHistory, null, 2));
-  } catch { /* best-effort */ }
-}
-
-/**
- * Record a notification, broadcast via SSE, and persist.
- * Called alongside fireEmailAlert / firePushAlert at every trigger point.
- */
-function addNotification(event, payload) {
-  const notif = {
-    id: crypto.randomUUID(),
-    event,
-    surface: eventSurface(event),
-    title: payload.title,
-    body: payload.body,
-    level: payload.level || 'info',
-    project: payload.project || '',
-    ts: new Date().toISOString(),
-    read: false,
-  };
-  notifHistory.unshift(notif);
-  if (notifHistory.length > MAX_NOTIF_HISTORY) notifHistory.length = MAX_NOTIF_HISTORY;
-  broadcast('notification', notif);
-  saveNotifHistory();
-  return notif;
-}
-
-// Load history at server start
-loadNotifHistory();
 
 // ── Memory: 4-layer file contents ─────────────────────────────────────────────
 function getMemory(cwd = process.cwd()) {
